refactor: 분리했던 컨트롤러를 하나의 컨트롤러로 수정한다
