+++ src/main/java/controller/GameController.java
@@ -1,6 +1,12 @@
 package controller;
 
+import domain.CardShuffler;
+import domain.card.Card;
+import domain.card.CardRandomShuffler;
+import domain.card.Deck;
 import domain.game.GameManager;
+import domain.game.GameResult;
+import domain.game.Result;
 import domain.participant.Dealer;
 import domain.participant.Participant;
 import domain.participant.Participants;
@@ -8,13 +14,16 @@ import view.InputView;
 import view.OutputView;
 
 import java.util.List;
+import java.util.Map;
+import java.util.stream.IntStream;
 
 import static view.message.Message.BLACKJACK_MESSAGE;
 import static view.message.Message.BUST_MESSAGE;
 import static view.message.Message.DEALER_DRAW_MESSAGE;
 
-public class GameProcessController {
+public class GameController {
 
+    private static final int START_GIVEN_COUNT = 2;
     private static final int PARTICIPANT_GIVEN_COUNT = 1;
     private static final int PLAYER_ORDER_OFFSET = 1;
     private static final int DEALER_ORDER = 0;
@@ -22,19 +31,59 @@ public class GameProcessController {
     private final InputView inputView;
     private final OutputView outputView;
 
-    private GameProcessController(final InputView inputView, final OutputView outputView) {
+    public GameController(final InputView inputView, final OutputView outputView) {
         this.inputView = inputView;
         this.outputView = outputView;
     }
 
-    public static GameProcessController create(final InputView inputView, final OutputView outputView) {
-        return new GameProcessController(inputView, outputView);
-    }
-
-    public void start(final Participants participants, final GameManager gameManager) {
+    public void start() {
+        final Participants participants = makeParticipants();
+        final Deck deck = makeDeck();
+        final GameManager gameManager = GameManager.create(deck, participants);
+        handCards(participants, gameManager);
+        printParticipantCards(participants);
         drawPlayersCard(participants, gameManager);
         handleDealerCards(participants, gameManager);
-        displayGameResult(participants);
+        printGameResult(participants);
+        printFinalGameResult(participants);
+    }
+
+    private Participants makeParticipants() {
+        return inputView.getInputWithRetry(() -> {
+            List<String> participantNames = inputView.getParticipantNames();
+            return Participants.create(participantNames);
+        });
+    }
+
+    private static Deck makeDeck() {
+        final CardShuffler cardShuffler = new CardRandomShuffler();
+        return Deck.create(cardShuffler);
+    }
+
+    private void handCards(final Participants participants, final GameManager gameManager) {
+        final int participantSize = participants.size();
+        IntStream.range(0, participantSize)
+                .forEach(participantOrder -> gameManager.giveCards(participantOrder, START_GIVEN_COUNT));
+    }
+
+    private void printParticipantCards(final Participants participants) {
+        outputView.printParticipantMessage(participants);
+        printDealerCard(participants);
+        printPlayerCards(participants);
+    }
+
+    private void printDealerCard(final Participants participants) {
+        final Dealer dealer = (Dealer) participants.getDealer();
+        final Card dealerFirstCard = dealer.getFirstCard();
+        outputView.printDealerCard(dealer.getName(), dealerFirstCard);
+    }
+
+    private void printPlayerCards(final Participants participants) {
+        final List<Participant> players = participants.getPlayer();
+        for (Participant player : players) {
+            List<Card> playerCards = player.getCard();
+            outputView.printPlayerCard(player.getName(), playerCards);
+        }
     }
 
     private void drawPlayersCard(final Participants participants, final GameManager gameManager) {
@@ -101,8 +150,23 @@ public class GameProcessController {
         }
     }
 
-    private void displayGameResult(final Participants participants) {
-        final GameResultController gameResultController = GameResultController.create(outputView);
-        gameResultController.start(participants);
+    private void printGameResult(final Participants participants) {
+        final Participant dealer = participants.getDealer();
+        printParticipantCardResult(dealer);
+        final List<Participant> players = participants.getPlayer();
+        players.forEach(this::printParticipantCardResult);
+    }
+
+    private void printParticipantCardResult(final Participant participant) {
+        final List<Card> participantCards = participant.getCard();
+        final int participantScore = participant.calculateScore();
+        outputView.printCardResult(participant.getName(), participantCards, participantScore);
+    }
+
+    private void printFinalGameResult(final Participants participants) {
+        final GameResult gameResult = GameResult.create(participants);
+        final Map<String, Result> playerGameResults = gameResult.getPlayerGameResults();
+        final Participant dealer = participants.getDealer();
+        outputView.printFinalGameResult(dealer.getName(), playerGameResults);
     }
 }
